refactored code to use CommitCard component
